Update 9-openstack-autoscaling-with-heat-ceilometer-gnocchi.rst
Fixing some grammar issues in the first part and then in the second part the results of the two nova lists are the same, which means that not autoscaling is shown. I'm not sure if deleting the second instance in the first table is right, but it gives the desired effect.